split and cleanup sha2
